Add HPS octave guard, phase period refinement, and SNR-grid corpus
DSP round 07 (benchmark-07-dsp):
- dsp/hps.rs: harmonic product spectrum guard against octave errors,
  consulted only when Hz probes are inconclusive; skipped entirely at
  confidence >= 0.97 so steady-state cost is zero
- dsp/phase.rs: phase-difference period refinement over two Hann windows
  (N/4 shift), magnitude^2-weighted, gated on confidence, stationarity
  and dominant fundamental; synthetic E6 error 0.406 -> 0.000 cents
- quality/noise.rs + example: deterministic seeded white-noise SNR grid
  (30/20/10 dB), report schema v2, per-level --check thresholds; clean
  gate unchanged

Benchmark: clean 19/19, mean MAE 2.645 -> 2.436 cents (-7.9%),
mean p95 4.945 -> 4.763; SNR grid 57/57 passed.

Also includes accumulated earlier-round work: biquad preprocessing,
generated gate thresholds (scripts/generate-gate-thresholds.mjs,
web/src/generated/gateThresholds.ts), quiet-reset loop handling,
default frequency range test, and benchmark reports 01-06.

+++ pitch-core/src/dsp/mpm.rs
@@ -35,17 +35,19 @@ impl MpmDetector {
 
         self.normalized_square_difference.resize(max_tau, 0.0);
         let calculation_start = min_tau.saturating_sub(1);
+        // Same rationale as YIN: accumulate the NSDF sums in f64 so long
+        // windows do not lose precision near the peak-picking thresholds.
         for tau in calculation_start..max_tau {
-            let mut numerator = 0.0;
-            let mut denominator = 0.0;
+            let mut numerator = 0.0_f64;
+            let mut denominator = 0.0_f64;
             for index in 0..(length - tau) {
-                let left = buffer[index];
-                let right = buffer[index + tau];
+                let left = f64::from(buffer[index]);
+                let right = f64::from(buffer[index + tau]);
                 numerator += left * right;
                 denominator += left * left + right * right;
             }
             self.normalized_square_difference[tau] = if denominator > 0.0 {
-                2.0 * numerator / denominator
+                (2.0 * numerator / denominator) as f32
             } else {
                 0.0
             };
